+++ src/app/cart/cart.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
 
 import { SharedModule } from '../shared/shared.module';
 
@@ -11,6 +12,7 @@ import { CartListComponent } from './components/cart-list/cart-list.component';
 @NgModule({
   imports: [
     CommonModule,
+    FormsModule,
     SharedModule
   ],
   declarations: [