+++ src/app/cart/model/cart.model.ts
@@ -1,8 +1,8 @@
-import { CartItem } from './cart-item.model';
+import { Product } from '../../product/model/product.model';
 
 export class Cart {
     constructor(
-        public items: CartItem[] = [],
+        public items: Product[] = [],
         public totalQuantity = 0,
         public totalPrice = 0
     ) {}