+++ src/app/product/components/product-list/product-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 
-import { Subscription } from 'rxjs';
+import { Subscription, Subject } from 'rxjs';
 
 import { Book } from '../../model/book.model';
 import { Product } from '../../model/product.model';
@@ -17,6 +17,7 @@ import { Cart } from '../../../cart/model/cart.model';
 })
 export class ProductListComponent implements OnInit, OnDestroy {
   books: Book[];
+  booksPromise: Promise<Book[]>;
 
   private sub: Subscription;
 
@@ -26,31 +27,32 @@ export class ProductListComponent implements OnInit, OnDestroy {
   ) {}
 
   ngOnInit() {
-    this.sub = this.productService.books$.subscribe((books: Book[]) => this.books = books);
-    this.sub.add(
-      this.cartService.cart$.subscribe((cart: Cart) => {
-        // Restore quantities if cart is empty
-        this.books.forEach((book: Book) => {
-          const isInCart = cart.items.length > 0 && cart.items.find(item => item.product.name === book.name);
+    // this.sub = this.productService.books$.subscribe((books: Book[]) => this.books = books);
+    // this.sub.add(
+    //   this.cartService.cart$.subscribe((cart: Cart) => {
+    //     // Restore quantities if cart is empty
+    //     this.books.forEach((book: Book) => {
+    //       const isInCart = cart.items.length > 0 && cart.items.find(item => item.product.name === book.name);
 
-          if (!isInCart) {
-            book.availableQuantity = book.quantity;
-          }
-        });
+    //       if (!isInCart) {
+    //         book.availableQuantity = book.quantity;
+    //       }
+    //     });
 
-        cart.items.forEach(cartItem => {
-          const bookIndex = this.books.findIndex(book => book.name === cartItem.product.name);
+    //     cart.items.forEach(cartItem => {
+    //       const bookIndex = this.books.findIndex(book => book.name === cartItem.product.name);
 
-          // Update available products quantity
-          if (bookIndex !== -1) {
-            const book = this.books[bookIndex];
-            book.availableQuantity = book.quantity - cartItem.quantity;
-          }
-        });
-      })
-    );
+    //       // Update available products quantity
+    //       if (bookIndex !== -1) {
+    //         const book = this.books[bookIndex];
+    //         book.availableQuantity = book.quantity - cartItem.quantity;
+    //       }
+    //     });
+    //   })
+    // );
 
-    this.productService.getProducts();
+    // this.productService.getProducts();
+    this.booksPromise = this.productService.getProductsPromise();
   }
 
   ngOnDestroy() {