+++ src/app/shared/shared.module.ts
@@ -2,12 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { HighlightDirective } from './directives/highlight.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 @NgModule({
   imports: [
     CommonModule
   ],
-  declarations: [HighlightDirective],
-  exports: [HighlightDirective]
+  declarations: [HighlightDirective, OrderByPipe],
+  exports: [HighlightDirective, OrderByPipe]
 })
 export class SharedModule { }